Make AdjacencyList a trait; first step to spec testing.
